refactor(app): extract click/fold geometry helpers to own module (phase 1g)
Three functions in input.rs — screen_to_buffer_position,
adjust_content_rect_for_compose, fold_toggle_byte_from_position — were
already written as associated functions (no `&self`, called via
`Self::foo(…)`). They're pure coordinate/geometry logic that only happens
to live on `Editor` by convention.

Move them verbatim to a new `app/click_geometry` module as free functions,
and rewrite the ~12 existing call sites across input.rs, mouse_input.rs,
and theme_inspect.rs from `Self::foo(…)` to
`super::click_geometry::foo(…)`. Drops input.rs by ~188 lines.

Add 7 unit tests covering the compose-rect centering math (no-op, odd
slack, min-width clamp) and the fallback/gutter rejection paths of
screen_to_buffer_position. These run without any Editor construction.

Part of phase 1 (free pure helpers) of
docs/internal/editor-modules-refactor-plan.md.
